fix: Stype StrEnum
